# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Player Team",
        "description": "Player team api documenatation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Player Team Api"
        }
    ],
    "paths": {
        "/api/v1": {},
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "description": "Returns the token of authenticated user",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Register"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvalidValidation"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login the user",
                "description": "Returns the token of authenticated user",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Login"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvalidCredentials"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout the user",
                "description": "Logout the user and revoke the token",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Logout"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "RegisterUserRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "passwordConfirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "passwordConfirmation": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ApiErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "example": "[]"
                    }
                },
                "type": "object"
            },
            "ApiSuccessResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "example": "[]"
                    }
                },
                "type": "object"
            },
            "Login": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "User is logged in successfully"
                            },
                            "data": {
                                "type": "object",
                                "example": {
                                    "token": "userToken"
                                }
                            }
                        }
                    }
                ]
            },
            "Logout": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "User is logged out successfully"
                            }
                        }
                    }
                ]
            },
            "Register": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "User is registered successfully"
                            },
                            "data": {
                                "type": "object",
                                "example": {
                                    "token": "userToken"
                                }
                            }
                        }
                    }
                ]
            },
            "InvalidCredentials": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiErrorResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "Your email or password is incorrect"
                            }
                        }
                    }
                ]
            },
            "InvalidValidation": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiErrorResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "Invalid value for :attribute"
                            }
                        }
                    }
                ]
            },
            "Unauthenticated": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiErrorResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "Unauthenticated"
                            }
                        }
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Player Team",
            "description": "Api endponts of app"
        }
    ]
}